requirements.txt
numpy>=1.21.0
opencv-python>=4.5.3
tensorflow>=2.8.0
matplotlib>=3.4.3
pandas>=1.3.0
tqdm>=4.62.3
ipython>=7.28.0
sympy>=1.9
streamlit>=1.2.0
scikit-learn>=0.24.2
pillow>=8.3.2
pytest==7.4.2
requests==2.31.0
kaggle==1.5.16 